Update work package

+++ config/database.sql
@@ -176,7 +176,7 @@ CREATE TABLE `tl_li_work_package` (
   `toProject` int(10) unsigned NOT NULL default '0',
   `toHourlyWage` int(10) unsigned NOT NULL default '0',
   `title` varchar(255) NOT NULL default '',
-  `hourLimit` int(10) unsigned NOT NULL default '0',
+  `hourLimit` double unsigned NOT NULL default '0',
   `isExternal` char(1) NOT NULL default '',
   PRIMARY KEY  (`id`),
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
